==> 04_spatial_analysis.Rmd <==
---
title: "Spatial Analysis"
author: "Brian Norman Peña-Calero"
date: "`r Sys.Date()`"
output: 
  rmdformats::downcute:
    self_contained: true
    highlight: kate
    toc_depth: 3
    default_style: dark
    code_folding: hide
    code_download: true
    highlight_downlit: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(sf)
library(car)
library(ape) # Analyses of Phylogenetics and Evolution
library(pgirmess)
library(leaflet)
# library(rgdal)
# library(raster)
# library(ggplot2)
# library(spatstat)
# library(plotrix)
# library(fields)
# library(leaflet)
# library(maptools)
# library(RColorBrewer)
# library(lattice)
# library(geoR)
# library(plotrix) 
# library(car)  # contains a function for logistic transformation (log odds) to make more normal

```

```{r}
# library(sp)
# 
# # Moran's I and spatial dependencies
# library(spdep) # Spatial Dependence: Weighting Schemes, Statistics and Models
# 
# library(pgirmess) # Data Analysis in Ecology
# 
# # Libraries for point processes
# library(spatstat)
# library(splancs) # K-function
# library(smacpod) # Spatial scanning statistic

```

# Import data

```{r}
ffi_total <- readRDS("01_data/processed/ffi_total.rds")
ffi_household <- read_csv("./01_data/raw/household_gps_share_072022.csv")
```


```{r}
ffi_prevalence_by_household <- ffi_total %>% 
  mutate(
    across(
      c(pf_recent:pv_historic, 
        recent_exposure, historical_exposure),
      ~ case_when(
        . == "Positive" ~ 1,
        TRUE ~ 0
      )
    )
  ) %>% 
  group_by(
    ffi_is_district,
    ffi_is_health_facility_name,
    ffi_is_community, 
    ffi_is_cod_com,
    ffi_is_cod_household
  ) %>% 
  summarise(
    examined = n(),
    across(
      c(pf_recent:pv_historic, 
        recent_exposure, historical_exposure),
      sum
    )
  ) %>% 
  ungroup() %>% 
  mutate(
    across(
      c(pf_recent:pv_historic, 
        recent_exposure, historical_exposure),
      ~ ./examined,
      .names = "prevalence_{.col}"
    )
  ) %>% 
  left_join(
    ffi_household %>% 
      select(ffi_is_community = ffi_h_community,
             ffi_is_cod_household = ffi_h_code_household,
             ffi_gps_long, ffi_gps_lat)
  ) %>% 
  mutate(
    cod_id = paste0(ffi_is_cod_com, ffi_is_cod_household)
  )
  # st_as_sf(
  #   coords = c("ffi_gps_long", "ffi_gps_lat"),
  #   crs = 4326
  # )


# ffi_prevalence_by_communities <- ffi_total %>% 
#   mutate(
#     recent_exposure = case_when(
#       recent_exposure == "Positive" ~ 1,
#       TRUE ~ 0
#     )
#   ) %>% 
#   group_by(ffi_is_community) %>% 
#   summarise(
#     examined = n(),
#     positives = sum(recent_exposure)
#   ) %>% 
#   ungroup() %>% 
#   mutate(
#     prevalence = positives/examined
#   )
```


```{r}
ffi_prevalence_by_household$log_odds <- logit(ffi_prevalence_by_household$prevalence_recent_exposure)
hist(ffi_prevalence_by_household$log_odds, xlab = "Log odds", main = "")
```

```{r}
pal = colorNumeric("Oranges", ffi_prevalence_by_household$prevalence_recent_exposure)
leaflet(ffi_prevalence_by_household) %>% 
  addTiles() %>% 
  addCircleMarkers(
    ~ffi_gps_long, ~ffi_gps_lat, fillOpacity=1,
    fillColor= ~pal(prevalence_recent_exposure),
    radius=~prevalence_recent_exposure*10,
    stroke=TRUE, weight=1
  ) %>% 
  addLegend(pal = pal, values = ~prevalence_recent_exposure)
```

## Global spatial autocorrelation

```{r}
hist(ffi_prevalence_by_household$prevalence_recent_exposure, xlab = "prevalence_recent_exposure", main = "")
```

```{r}
ffi.dists <- as.matrix(dist(cbind(ffi_prevalence_by_household$ffi_gps_long, 
                                  ffi_prevalence_by_household$ffi_gps_lat)))
dim(ffi.dists) # 980 x 980 matrix of distance between all sets of points
```

### First aproach

```{r error=TRUE}
# Take the inverse of the matrix values so that closer values have a larger weight and vs vs
ffi.dists.inv <- 1/ffi.dists
diag(ffi.dists.inv) <- 0   # replace the diagonal values with zero

# Computes Moran's I autocorrelation coefficient of x giving a matrix of weights (here based on distance) 
Moran.I(ffi_prevalence_by_household$log_odds, ffi.dists.inv)                 # from the "ape" package
```

### Second aproach
```{r}
maxDist<-max(dist(cbind(ffi_prevalence_by_household$ffi_gps_long, 
                        ffi_prevalence_by_household$ffi_gps_lat)))
maxDist
```



```{r dpi = 300, fig.height=7, fig.width=9}
xy=cbind(ffi_prevalence_by_household$ffi_gps_long,
         ffi_prevalence_by_household$ffi_gps_lat)
pgi.cor <- correlog(coords=xy, 
                    z=ffi_prevalence_by_household$log_odds,
                    method="Moran", nbclass=10)   # "pgirmess" package
# coords = xy cordinates, z= vector of values at each location and nbclass = the number of bins
plot(pgi.cor) # statistically significant values (p<0.05) are plotted in red
```

```{r}
pgi.cor
```

Based on the correlogram, over what spatial lags are there evidence for spatial autocorrelation? Is this clustering positive or negative?

Compare the correlogram to the results from a semivariogram approach:

```{r}
ffi_malaria_data_geo <- geoR::as.geodata(
  ffi_prevalence_by_household[,c("ffi_gps_long","ffi_gps_lat","log_odds")]
)

# Generate and plot a binned variogram (10 bins) NB: have made for full max distance (even though likely inaccurate) for comparison
Vario<- geoR::variog(ffi_malaria_data_geo,
                     max.dist = maxDist,
                     uvec = seq(first(pgi.cor[, 1]),
                              last(pgi.cor[, 1]),
                              l = 10))
```


```{r dpi = 400, fig.height=10, fig.width=7}
par(mfrow=c(2,1))
plot(Vario)
plot(pgi.cor)
```


## Another approach

```{r}
library(spdep)
```


```{r}
# Problema de mismas ubicaciones
# ffi_prevalence_by_household %>% 
#   add_count(ffi_gps_long, ffi_gps_lat) %>% 
#   filter(n >1) %>% 
#   View()

ffi_prevalence_by_household_sf <- ffi_prevalence_by_household %>% 
  st_as_sf(
    coords = c("ffi_gps_long", "ffi_gps_lat"),
    crs = 4326
  ) 
  

ffi_prevalence_kmeans <- ffi_prevalence_by_household_sf %>% 
  group_nest(
    ffi_is_district,
    #ffi_is_community
    #ffi_is_health_facility_name
  ) %>% 
  mutate(
    ffi_coords = map(
      data, 
      ~ coordinates(as(., "Spatial"))
    ),
    neigh_nb = map2(
      data, ffi_coords,
      ~ knn2nb(
          knearneigh(
            .y,
            k = 1,
            longlat = TRUE
          ),
          row.names = .x$cod_id
        )
    ),
    neigh_kd1 = pmap(
      list(ffi_coords,
           neigh_nb,
           data),
      ~ dnearneigh(
        ..1,
        d1 = 0, 
        d2 = max(unlist(nbdists(..2, ..1))), 
        row.names = ..3$cod_id
      )
    ),
    self = map(
      neigh_kd1,
      ~ nb2listw(
        .,
        style = "W", 
        zero.policy = TRUE
      )
    )
  )

# 
# ffi_prevalence_by_household_sf <- ffi_prevalence_by_household_sf %>% 
#   filter(ffi_is_community == "3 DE OCTUBRE")
# 
# ffi_prevalence_coords <- as(ffi_prevalence_by_household_sf, "Spatial") %>%
#   coordinates()
# 
# ffi_ids <- ffi_prevalence_by_household_sf$cod_id
# 
# 
# Neigh_nb <- knn2nb(
#   knearneigh(ffi_prevalence_by_household_sf,
#              k = 1),
#   row.names = ffi_ids
# )
# 
# dsts <- unlist(nbdists(Neigh_nb, ffi_prevalence_coords))
# max_1nn <- max(dsts)
# Neigh_kd1 <- dnearneigh(ffi_prevalence_coords,
#                         d1 = 0,
#                         d2 = max_1nn,
#                         row.names = ffi_ids)
# 
# self <- nb2listw(Neigh_kd1,
#                  style="W",
#                  zero.policy = T)
```

```{r}

breaks <- c(-Inf, -2.58, -1.96, 
            -1.65, 1.65, 1.96,
            2.58, Inf)
labels <- c("Cold spot: 99% confidence", 
            "Cold spot: 95% confidence", 
            "Cold spot: 90% confidence", 
            "PSU Not significant",
            "Hot spot: 90% confidence",
            "Hot spot: 95% confidence",
            "Hot spot: 99% confidence")

ffi_w.getis <- ffi_prevalence_kmeans %>% 
  mutate(
    data_nosf = map(
      data,
      ~ .x %>% 
        st_set_geometry(NULL) %>% 
        pivot_longer(
          cols = c(prevalence_pf_recent:prevalence_historical_exposure),
          names_to = "type_plasmodium",
          values_to = "prevalence"
        ) %>% 
        group_nest(type_plasmodium,
                   .key = "data_nosf")
    )
  ) %>% 
  unnest(data_nosf) %>% 
  mutate(
    LISA = map2(data_nosf,
               self,
               ~ localG(.x$prevalence, .y)),
    LISA_clust = map(LISA,
                     ~ cut(., include.lowest = TRUE,
                           breaks = breaks, 
                           labels = labels)),
    LISA_label_orig = map(LISA,
                          ~ attributes(.)$cluster),
    LISA = map(LISA, as.numeric)
  ) %>% 
  select(
    ffi_is_district,
    #ffi_is_community,
    #ffi_is_health_facility_name,
    type_plasmodium:LISA_label_orig
  ) %>% 
  unnest(cols = c(data_nosf, LISA, 
                  LISA_clust, LISA_label_orig))



# w.getis <- ffi_prevalence_by_household_sf %>%
#   st_set_geometry(NULL) %>%
#   pivot_longer(
#     cols = c(prevalence_pf_recent:prevalence_historical_exposure),
#     names_to = "type_plasmodium",
#     values_to = "prevalence"
#   ) %>%
#   group_nest(type_plasmodium) %>%
#   mutate(
#     LISA = map(data,
#                ~ localG(.$prevalence, self, return_internals = TRUE)),
#     LISA_clust = map(LISA,
#                      ~ cut(., include.lowest = TRUE,
#                            breaks = breaks,
#                            labels = labels)),
#     LISA_label_orig = map(LISA,
#                           ~ attributes(.)$cluster),
#     LISA = map(LISA, as.numeric)
#   ) %>%
#   unnest() %>% 
#   pull(LISA_label_orig)
# 
# c <- ffi_prevalence_by_household_sf %>%
#   st_set_geometry(NULL) %>%
#   pivot_longer(
#     cols = c(prevalence_pf_recent:prevalence_historical_exposure),
#     names_to = "type_plasmodium",
#     values_to = "prevalence"
#   )
# 
# cut(c$prevalence, c(-Inf, mean(c$prevalence), Inf), labels = c("Low", 
#     "High"))
```


```{r}
ffi_w.getis <- ffi_w.getis %>% 
  left_join(
    ffi_household %>% 
      select(
        ffi_is_district = ffi_h_district,
        ffi_is_cod_com = ffi_h_code_community,
        ffi_is_cod_household = ffi_h_code_household,
        ffi_gps_lat, 
        ffi_gps_long
      )
  )

w.getis_sf <- ffi_w.getis %>% 
  st_as_sf(coords = c("ffi_gps_long", 
                      "ffi_gps_lat"),
           crs = 4326,
           remove = FALSE)

# w.getis_sf
# 
# ffi_w.getis %>%
#   filter(type_plasmodium %in% c("prevalence_pf_recent",
#                                           "prevalence_pv_recent",
#                                           "prevalence_recent_exposure")) %>%
#   select(1:5, prevalence, LISA, LISA_clust, LISA_label_orig) %>%
#   View()
#   hacksaw::keep_na(.logic = "OR", ) %>%
#   count(LISA_label_orig)
#   print(n = Inf)
```


```{r}
library(gtsummary)

fisher.test.simulate.p.values <- function(data, variable, by, ...) {
  result <- list()
  test_results <- stats::fisher.test(data[[variable]], data[[by]], simulate.p.value = TRUE)
  result$p <- test_results$p.value
  result$test <- test_results$method
  result
}
```

```{r}

ffi_w.getis

labelled::var_label(ffi_w.getis) <- list(
  prevalence_historical_exposure = "Historical Exposure",
  prevalence_pf_historic = "P. falciparum historical",
  prevalence_pf_recent = "P. falciparum recent",
  prevalence_pv_historic  = "P. vivax historical",
  prevalence_pv_recent = "P. vivax recent",
  prevalence_recent_exposure = "Recent Exposure"
)

ffi_w.getis <- ffi_w.getis %>% 
  mutate(
    type_plasmodium = factor(type_plasmodium,
                             levels = c("prevalence_historical_exposure",
                                        "prevalence_pf_historic",
                                        "prevalence_pv_historic",
                                        "prevalence_recent_exposure",
                                        "prevalence_pf_recent",
                                        "prevalence_pv_recent"),
                             labels = c("Historical Exposure",
                                        "P. falciparum historical",
                                        "P. vivax historical",
                                        "Recent Exposure",
                                        "P. falciparum recent",
                                        "P. vivax recent"))
  )


tab_getis_strata <- ffi_w.getis %>%
  select(ffi_is_district, type_plasmodium, LISA_clust) %>% 
  # pivot_wider(
  #   names_from = type_plasmodium,
  #   values_from = LISA_clust
  # ) %>% 
  # select("prevalence_pf_recent", 
  #        "prevalence_pv_recent",
  #        "prevalence_recent_exposure") %>% 
  tbl_strata(
  strata = type_plasmodium,
  .tbl_fun =
    ~ .x %>%
        tbl_summary(
          by = ffi_is_district,
          digits = everything() ~ c(0, 1)
        ) %>%
        add_p(
          test = list(all_categorical() ~ "fisher.test.simulate.p.values"),
          pvalue_fun = scales::pvalue
        ) %>%
        bold_p() %>%
        modify_header(label = "**Variable**") %>%
        bold_labels() 
  ) 
  
tab_getis_strata %>%
  as_flex_table() %>%
  flextable::save_as_docx(path = "./02_output/reports/tab_getis_strata.docx")
  
tab_getis_general <- ffi_w.getis %>%
  select(type_plasmodium, LISA_clust) %>% 
  gtsummary::tbl_summary(
    by = type_plasmodium,
    digits = everything() ~ c(0, 1)
  ) %>% 
  add_n() %>%
  add_overall(
    last = TRUE,
    digits = everything() ~ c(0, 1)
  ) %>%
  add_p(
     test = list(all_categorical() ~ "fisher.test.simulate.p.values"),
     pvalue_fun = scales::pvalue
  ) %>% 
  bold_p() %>%
  modify_header(label = "**Variable**") %>%
  bold_labels() %>% 
  modify_spanning_header(stat_0 ~ "**Overall**")

tab_getis_general %>%
  as_flex_table() %>%
  flextable::save_as_docx(path = "./02_output/reports/tab_getis_general.docx")

```


```{r}
data("Peru", package = "innovar")

ffi_shp <- Peru %>% 
  filter(dep == "LORETO",
         distr %in% c("INDIANA",
                      "BELEN"))

plot_getis_recent <- w.getis_sf %>% 
  filter(type_plasmodium %in% c("prevalence_pf_recent", 
                                "prevalence_pv_recent",
                                "prevalence_recent_exposure")) %>% 
  mutate(type_plasmodium = fct_relevel(type_plasmodium,
                                       "prevalence_recent_exposure")) %>% 
  ggplot() +
  geom_sf(data = ffi_shp, fill = "#d9ead8") + 
  geom_sf(aes(col = LISA_clust), 
          size = 1.5,
          alpha = 1) +
  geom_sf(data = ffi_shp, fill = NA) + 
  innovar::scale_color_innova("npr", na.value = "grey50") +
  # colorspace::scale_color_discrete_diverging(palette = "Blue-Red",
  #                                limits = labels) +
  facet_wrap(vars(type_plasmodium),
             labeller = labeller(type_plasmodium = c(prevalence_recent_exposure = "Prevalence Recent Exposure",
                                   prevalence_pf_recent = "Prevalence Recent P. Falciparum",
                                   prevalence_pv_recent = "Prevalence Recent P. Vivax"))) +
  labs(color = "Local Indicator of Spatial Association",
       title = "Spatial Analysis for Recent Exposure") +
  #guides(colour = guide_legend(override.aes = list(size=10))) +
  theme_classic(base_size = 12) +
  theme(
    legend.title = element_text(
      face = "bold"
    ),
    plot.title = element_text(
      face = "bold",
      hjust = 0.5
    ),
    strip.text = element_text(
      face = "bold"
    )
  )
plot_getis_recent
```

```{r eval = FALSE}
ggsave("02_output/plots/map_getis_lisa_recent.png",
       plot_getis_recent,
       dpi = 400,
       scale = 1.2,
       device = grDevices::png)
```

```{r}
plot_getis_historical <- w.getis_sf %>% 
  filter(type_plasmodium %in% c("prevalence_pf_historic", 
                                "prevalence_pv_historic",
                                "prevalence_historical_exposure")) %>% 
  ggplot() +
  geom_sf(data = ffi_shp, fill = "#d9ead8") + 
  geom_sf(aes(col = LISA_clust), 
          size = 1.5,
          alpha = 1) +
  geom_sf(data = ffi_shp, fill = NA) + 
  innovar::scale_color_innova("npr", na.value = "grey50") +
  # colorspace::scale_color_discrete_diverging(palette = "Blue-Red",
  #                                limits = labels) +
  facet_wrap(vars(type_plasmodium),
             labeller = labeller(type_plasmodium = c(prevalence_historical_exposure = "Prevalence Historic Exposure",
                                   prevalence_pf_historic = "Prevalence Historic P. Falciparum",
                                   prevalence_pv_historic = "Prevalence Historic P. Vivax"))) +
  labs(color = "Local Indicator of Spatial Association",
       title = "Spatial Analysis for Historic Exposure") +
  #guides(colour = guide_legend(override.aes = list(size=10))) +
  theme_classic(base_size = 12) +
  theme(
    legend.title = element_text(
      face = "bold"
    ),
    plot.title = element_text(
      face = "bold",
      hjust = 0.5
    ),
    strip.text = element_text(
      face = "bold"
    )
  )
plot_getis_historical
```

```{r eval = FALSE}
ggsave("02_output/plots/map_getis_lisa_historical.png",
       plot_getis_historical,
       dpi = 400,
       scale = 1.2,
       device = grDevices::png)
```


```{r}
# resI <- localmoran(ffi_prevalence_by_household_sf$prevalence_exposure,
#                    nb2listw(Neigh_kd1))
# 
# printCoefmat(data.frame(resI),
#              check.names=FALSE) %>% 
#   as_tibble() %>% 
#   mutate(
#     ffi_is_community = ffi_prevalence_by_household_sf$ffi_is_community,
#     ffi_is_cod_household = ffi_prevalence_by_household_sf$ffi_is_cod_household,
#     .before = 1
#   )
```
## Another analysis

```{r}
# ffi_prevalence_by_household %>% 
#   st_as_sf(
#     coords = c("ffi_gps_long", "ffi_gps_lat"),
#     crs = 4326
#   ) %>% 
#   tm_shape() +
#   tm_fill("positives", style = "jenks", n = 6) +
#   tm_borders() +
#   tm_layout(legend.outside = TRUE, legend.outside.position = "left")
```
